chore(batcher): refactor driver tests
Move large driver integration tests from the inline cfg(test) block into
crates/batcher/core/tests/driver.rs, gating test_utils sub-modules behind
cfg(any(test, feature = "test-utils")) so integration tests can access
shared fixtures without redefining stubs. Replace free-floating helper
functions with SubmissionStub and DriverFixture unit structs in
test_utils/builder.rs per CLAUDE.md conventions. Convert throttle unit
tests to rstest parameterized cases. Expose DriverEvent publicly.
