+++ crates/batcher/core/src/test_utils/mod.rs
@@ -2,3 +2,21 @@
 
 mod throttle;
 pub use throttle::{ThrottleCallLog, TrackingThrottleClient};
+
+mod pipeline;
+pub use pipeline::{OneReorgPipeline, Recorded, ReorgPipeline, TrackingPipeline};
+
+#[cfg(any(test, feature = "test-utils"))]
+mod source;
+#[cfg(any(test, feature = "test-utils"))]
+pub use source::{OneBlockSource, PendingL1HeadSource, PendingSource};
+
+#[cfg(any(test, feature = "test-utils"))]
+mod builder;
+#[cfg(any(test, feature = "test-utils"))]
+pub use builder::{DriverFixture, SubmissionStub};
+
+#[cfg(any(test, feature = "test-utils"))]
+mod tx_manager;
+#[cfg(any(test, feature = "test-utils"))]
+pub use tx_manager::{ImmediateConfirmTxManager, ImmediateFailTxManager, NeverConfirmTxManager};
